terminado

+++ LAB2-2.c
@@ -1,7 +1,7 @@
 /*
 Implementar un programa en C que informe por pantalla si una cadena,introducida
 como argumento en la ejecución del programa, es un palíndromo(una cadena es un palíndromo
- si se lee igual de izquierda a derecha que dederecha a izquierda).
+si se lee igual de izquierda a derecha que dederecha a izquierda).
 
 */
 #include <stdio.h>
@@ -18,7 +18,7 @@ int main(int argc, char *argv[])
 
    // Obtener la cadena del argumento
    char *cadena = argv[1];
-   int longitud = strlen(cadena);
+   size_t longitud = strlen(cadena);
 
    // Comprobar si la cadena es un palíndromo
    int es_palindromo = 1;
